+++ src/app/subscription/layout.tsx
@@ -1,10 +1,17 @@
 import { TypeChildren } from "@/types/common";
 import dynamic from "next/dynamic";
+import { Suspense } from "react";
 
 const ClientLayout = dynamic(() => import("@/components/layout/ClientLayout"), {
   ssr: false,
 });
 
 export default function Layout({ children }: TypeChildren) {
-  return <ClientLayout>{children}</ClientLayout>;
+  return (
+    <ClientLayout>
+      <Suspense fallback={<div>페이지를 불러오는 중입니다</div>}>
+        {children}
+      </Suspense>
+    </ClientLayout>
+  );
 }
